refactor: make Graph component modular
